+++ wezterm-gui/src/termwindow/selection.rs
@@ -3,6 +3,7 @@ use ::window::WindowOps;
 use mux::pane::{Pane, PaneId};
 use std::cell::RefMut;
 use std::rc::Rc;
+use termwiz::surface::Line;
 use wezterm_term::StableRowIndex;
 
 impl super::TermWindow {
@@ -10,6 +11,59 @@ impl super::TermWindow {
         RefMut::map(self.pane_state(pane_id), |state| &mut state.selection)
     }
 
+    /// Returns the selection region as a series of Line
+    pub fn selection_lines(&self, pane: &Rc<dyn Pane>) -> Vec<Line> {
+        let mut result = vec![];
+
+        let rectangular = self.selection(pane.pane_id()).rectangular;
+        if let Some(sel) = self
+            .selection(pane.pane_id())
+            .range
+            .as_ref()
+            .map(|r| r.normalize())
+        {
+            let mut last_was_wrapped = false;
+            let first_row = sel.rows().start;
+            let last_row = sel.rows().end;
+
+            for line in pane.get_logical_lines(sel.rows()) {
+                if result.is_empty() || !last_was_wrapped {
+                    result.push(Line::with_width(0, line.physical_lines[0].current_seqno()));
+                }
+                let last_idx = line.physical_lines.len().saturating_sub(1);
+                for (idx, phys) in line.physical_lines.iter().enumerate() {
+                    let this_row = line.first_row + idx as StableRowIndex;
+                    if this_row >= first_row && this_row < last_row {
+                        let last_phys_idx = phys.cells().len().saturating_sub(1);
+                        let cols = sel.cols_for_row(this_row, rectangular);
+                        let last_col_idx = cols.end.saturating_sub(1).min(last_phys_idx);
+                        let mut col_span = phys.columns_as_line(cols);
+                        let seqno = col_span.current_seqno();
+                        // Only trim trailing whitespace if we are the last line
+                        // in a wrapped sequence
+                        if idx == last_idx {
+                            col_span.prune_trailing_blanks(seqno);
+                        }
+
+                        result
+                            .last_mut()
+                            .map(|line| line.append_line(col_span, seqno));
+
+                        last_was_wrapped = last_col_idx == last_phys_idx
+                            && phys
+                                .cells()
+                                .get(last_col_idx)
+                                .map(|c| c.attrs().wrapped())
+                                .unwrap_or(false);
+                    }
+                }
+            }
+        }
+
+        result
+    }
+
+    /// Returns the selection text only
     pub fn selection_text(&self, pane: &Rc<dyn Pane>) -> String {
         let mut s = String::new();
         let rectangular = self.selection(pane.pane_id()).rectangular;